Better error handling for shrink and fixes for some other error cases
* If shrink fails, a comprehensible error message is shown to the user, recommending to not use shrinking anymore for this game.
* Cleanup directories with _ on the SD card before making copies.

+++ GDEmuSdCardManager.BLL/SdCardManager.cs
@@ -127,7 +127,11 @@ namespace GDEmuSdCardManager.BLL
                 //    //throw new System.Exception("There was an error while shriking the GDI: " + commandResult.StandardError);
                 //}
 
-                var gdiPath = Directory.EnumerateFiles(destinationFolder).Single(f => Path.GetExtension(f) == ".gdi");
+                var gdiPath = Directory.EnumerateFiles(destinationFolder).SingleOrDefault(f => Path.GetExtension(f) == ".gdi");
+                if(gdiPath == null)
+                {
+                    throw new OperationCanceledException($"Could not shrink {game.GameName}. You might need to copy it without shrinking.");
+                }
                 var newGdi = GameManager.GetGdiFromFile(gdiPath);
                 File.Delete(gdiPath);
                 newGdi.SaveTo(Path.Combine(destinationFolder, "disc.gdi"), true);
